+++ CHANGELOG.md
@@ -1,9 +1,17 @@
-### [0.0.2]
+# [0.0.3]
+## Added:
+- Add `Gtk.Entry` and `Gtk.AboutDialog` Widgets.
+ > **Note:** `Gtk.Entry` properties are not fully added yet i will added this in `v0.0.4` after my exam I have exam now.
+- Add `AboutDialog` Properties.
+## Fixed:
+- Fix Readme Typo.
+
+# [0.0.2]
 ## Added:
 - Add `set_text` and `get_text` methods for Label widget.
 ## Fixed:
 - Improved logic for unified completion provider to prevent suggestion conflicts.
 
-### [0.0.1]
+# [0.0.1]
 
 - Initial release